requirements.txt
atlassian-python-api>=3.41.0
mcp>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.20.0
starlette>=0.20.0
